requirements.txt
Pillow
docopt
numpy
opencv-python
SpeechRecognition
